+++ renderer/src/renderer/image_loader.rs
@@ -184,7 +184,8 @@ impl PixelData {
             ktxvulkan_sys::ktxTexture2_CreateFromMemory(
                 cubemap_tex_data.as_ptr(),
                 cubemap_tex_data.len(),
-                ktxvulkan_sys::ktxTextureCreateFlagBits_KTX_TEXTURE_CREATE_LOAD_IMAGE_DATA_BIT,
+                // i32 on windows and u32 on linux :)
+                ktxvulkan_sys::ktxTextureCreateFlagBits_KTX_TEXTURE_CREATE_LOAD_IMAGE_DATA_BIT as u32,
                 &mut new_tex,
             )
         };
